Add whereAdded to by-name params for better errors
The whereAdded member was added to the ByName type. This is so that we
can find the other use of the name rather than just know that this use
is invalid.

A new SetterType method on the ByName type was added. This returns the
type name of the embedded setter and is useful when generating a zsh
argument completion function.

The FinalCheckFunc type definition is moved into the pSet.go file.

A progBaseName member was added to the PSet type

The PSet.nameCheck function now gives a more detailed error report
with extra details of where the parameter was set up from.