Remove the AVM and NilAVM types
The change is a breaking change so we also move everything to v4

The AValMap type (with an embedded AllowedValues) is replaced with just
the map of allowed values. This makes for easier setter initialisation
and a simpler interface.

The AllowedValuesMap method is removed from the Setter interface and a
new interface AllowedValuesMapper is added.

We no longer need to add the NilAVM type to all of our setters.