Regularise the help-printing functions
The different settings of the help style calls different help functions to
print the different parts of the message. These functions have been made to
all have the same signature and then put into a map so that the
StdHelp.Help() function can be simplified. It also makes it a bit more robust
as it will now report an unexpected style.

Also the standard help message has been made more concise.