Add a SeeAlso list to the ByName parameter
The programmer can now add references to other parameters. These references
must be:
1- unique - the same name may not appear twice in the see-also list for a
single param
2- valid - any name referenced must be the name of some parameter.

Any breach of these will cause a panic. The first is checked as the reference
is added and the second is checked before the parameters are parsed.

The SeeAlso method on the ByName type returns a slice of strings
The SeeAlso function returns an OptFunc which can be used to add new
SeeAlso references