Add conditional ActionFunc generators
This adds the new ParamTestFunc type that takes the same parameters as an
ActionFunc but returns a bool. This is then used to test conditions before
setting values in the generated action function. There are new functions
which use this new type:

SetBoolIf
SetStringIf
SetMapIf

There is also a new ActionFunc generator
SetMap