Use errutil.ErrMap in place of the local ErrMap
This allows a simplification and refactoring of the phelp error reporting
with a slight change in the format of the reports.

Many tests have to be changed to reflect the new output of the error
reporting and the change to the location.Err.Error() output.

The changes to the twrap.Wrap() function allow for some nicer help messages
so they have been changed to use that.