Show alternative param names in a more concise form
Also:

show the value type rather than just '...'. The value type is the name of the
Setter type with any package name stripped off and trailing digits
removed. If the Setter also satisfies the ValDescriber interface then the
result of calling that method is used instead.

don't show the initial value if it is a blank string, a false bool or a zero
number unless it has changed value