feature: change the group names for the common parameters
the names change from being prefixed with common.params to stdParams. This
makes the names shorter and so makes some of the help messages slightly
easier to read. It also makes the group config file for completion parameters
shorter. The downside is that the file name changes but then the path would
have changed anyway as the major version number is a part of the name.