feature(psetter): add the SuggestionString and SuggestedVals funcs
The SuggestionString func constructs a "did you mean..." string suitable to
be used by param.Setter when constructing an error message.

The SuggestedVals func is a standard func for constructing a slice of
alternative values suitable for passing to the SuggestionString func.

Also change the psetter.TimeLocation and psetter.Pathname setters to use
SuggestedVals and SuggestionString